+++ package.json
@@ -9,6 +9,7 @@
     "@testing-library/user-event": "^7.2.1",
     "bootstrap": "^4.5.2",
     "gatsby": "^4.8.1",
+    "gatsby-plugin-layout": "^3.9.0",
     "gatsby-plugin-sass": "^5.8.0",
     "jquery": "^3.6.0",
     "popper.js": "^1.16.1",
