+++ public/page-data/projects/page-data.json
@@ -1,0 +1,5 @@
+{
+    "componentChunkName": "component---src-pages-projects-js",
+    "path": "/projects/",
+    "result": {"pageContext":{}},
+    "staticQueryHashes": []}
