+++ src/pages/about.js
@@ -2,28 +2,24 @@ import React from 'react';
 
 import Hero from '../components/Hero';
 import Content from '../components/Content';
-import Project from '../components/Project';
 import { useSpring, animated } from 'react-spring';
 import data from '../data/data';
 
-function ProjectsPage(props) {
+function AboutPage() {
 
-    const page = props.page;
+    const page = data.pages[2];
 
-    const style = useSpring(data.fadeIn);
-    const projects = data.projects;
+    const style = useSpring(data.fadeIn)
 
     return(
         <animated.div style={style}>
             <Hero hero={page.hero} />
             <Content>
-                {
-                    projects.map(project => <Project project={project} />)
-                }
+                {data.about}
             </Content>
         </animated.div>  
     );
 
 }
 
-export default ProjectsPage;
+export default AboutPage;
